Refactor isis_archive_monitor into its own package
Refs #11
